diy-q-business-chatbot/app/requirements.txt
boto3
faiss-cpu
langchain
langchain_aws
langchain_community
pypdf
streamlit